Replace HTTP client and promise libraries
- request -> superagent
- when    -> bluebird

The motivation for replacing 'request' to make this SDK browser-compatible.
Superagent is an HTTP client library that can run in node or in the browser.

Reasons for replacing 'when' are twofold:

Since the last version of this library, the Promise specification has been
standardized; using a promise implementation that closely matches that
specification seems like a good idea.

Bluebird has become an exceedingly popular option for promises - even in cases
where a native `Promise` implementation is available.
Bluebird is fast - generally faster than native promise implementations;
it is compatible with the Promise specification;
and Bluebird comes with some useful extra features, such as enhanced error
reporting, and the `promisify` function to convert node-style async functions to
promise-returning functions.