Fix Promise Callback
Since switching from `when` to `Bluebird`, we've had a minor bug where callback parameters in a promise were assumed to be wrapped in an array. This could cause authentication by way of Passport to fail. Fix the issue by passing the parameters normally and treating them like the first-class citizens they really are.